# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Obsura API",
    "summary": "Review-first sanitization workflows, studio assets, and job orchestration for Obsura.",
    "description": "Obsura API is the review-first workflow and orchestration layer for Obsura.\n\nThe API supports:\n- persistent studio assets for custom detection and reusable configurations\n- reviewable jobs and run history\n- text-first and screenshot-first sanitization workflows\n- transformation behavior beyond plain placeholder replacement\n- image-region protection flows for blur, pixelation, masks, and overlays\n\nThis OpenAPI document is intended for direct import into API clients such as Postman.",
    "contact": {
      "name": "Obsura",
      "url": "https://github.com/obsura"
    },
    "license": {
      "name": "AGPL-3.0-or-later"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/api/v1/workflows/text/analyze": {
      "post": {
        "tags": [
          "text-workflows"
        ],
        "summary": "Analyze Text",
        "description": "Analyze text-like content into reviewable findings and optionally persist a job.",
        "operationId": "analyze_text_api_v1_workflows_text_analyze_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TextAnalysisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_TextAnalysisResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workflows/text/transform": {
      "post": {
        "tags": [
          "text-workflows"
        ],
        "summary": "Transform Text",
        "description": "Transform a persisted text job after review decisions and overrides are applied.",
        "operationId": "transform_text_api_v1_workflows_text_transform_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TextTransformRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_TextTransformResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/workflows/text/analyze-transform": {
      "post": {
        "tags": [
          "text-workflows"
        ],
        "summary": "Analyze And Transform Text",
        "description": "Run a one-shot analyze-plus-transform flow while still producing reviewable findings.",
        "operationId": "analyze_and_transform_text_api_v1_workflows_text_analyze_transform_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TextAnalysisRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse_TextTransformResponse_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiError": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ],
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ApiError",
        "description": "Unified error payload."
      },
...
      "ApiResponse_TextAnalysisResponse_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TextAnalysisResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "pagination": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PaginationMeta"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApiError"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "ApiResponse[TextAnalysisResponse]"
      },
      "ApiResponse_TextTransformResponse_": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "data": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TextTransformResponse"
              },
              {
                "type": "null"
              }
            ]
          },
          "pagination": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PaginationMeta"
              },
              {
                "type": "null"
              }
            ]
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ApiError"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "success"
        ],
        "title": "ApiResponse[TextTransformResponse]"
      },
...
      "BoundingBox": {
        "properties": {
          "x": {
            "type": "integer",
            "minimum": 0.0,
            "title": "X"
          },
          "y": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Y"
          },
          "width": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Width"
          },
          "height": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "Height"
          }
        },
        "type": "object",
        "required": [
          "x",
          "y",
          "width",
          "height"
        ],
        "title": "BoundingBox",
        "description": "A region on an image or screenshot."
      },
...
      "ContentType": {
        "type": "string",
        "enum": [
          "text",
          "structured_text",
          "screenshot",
          "image"
        ],
        "title": "ContentType"
      },
...
      "FindingKind": {
        "type": "string",
        "enum": [
          "text_span",
          "image_region",
          "face_region",
          "face_subregion"
        ],
        "title": "FindingKind"
      },
      "FindingOverride": {
        "properties": {
          "finding_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finding Id"
          },
          "start_index": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Index"
          },
          "end_index": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Index"
          },
          "decision": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReviewDecision"
              },
              {
                "type": "null"
              }
            ]
          },
          "transformation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TransformationRule"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "title": "FindingOverride",
        "description": "Override a stored or transient finding before transformation."
      },
      "FindingRecord": {
        "properties": {
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          },
          "job_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Id"
          },
          "source": {
            "$ref": "#/components/schemas/FindingSource"
          },
          "kind": {
            "$ref": "#/components/schemas/FindingKind"
          },
          "entity_type": {
            "type": "string",
            "title": "Entity Type"
          },
          "entity_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entity Name"
          },
          "start_index": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Index"
          },
          "end_index": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Index"
          },
          "region": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/BoundingBox"
              },
              {
                "type": "null"
              }
            ]
          },
          "matched_text_preview": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Matched Text Preview"
          },
          "matched_text_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Matched Text Hash"
          },
          "confidence": {
            "type": "number",
            "maximum": 1.0,
            "minimum": 0.0,
            "title": "Confidence",
            "default": 1.0
          },
          "decision": {
            "$ref": "#/components/schemas/ReviewDecision",
            "default": "pending"
          },
          "transformation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TransformationRule"
              },
              {
                "type": "null"
              }
            ]
          },
          "metadata": {
            "additionalProperties": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                },
                {
                  "type": "boolean"
                },
                {
                  "items": {
                    "type": "string"
                  },
                  "type": "array"
                }
              ]
            },
            "type": "object",
            "title": "Metadata"
          }
        },
        "type": "object",
        "required": [
          "source",
          "kind",
          "entity_type"
        ],
        "title": "FindingRecord",
        "description": "Reviewable finding for text or image workflows."
      },
      "FindingSource": {
        "type": "string",
        "enum": [
          "built_in",
          "custom",
          "manual",
          "face",
          "ocr"
        ],
        "title": "FindingSource"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ManualTextSpan": {
        "properties": {
          "start_index": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Start Index"
          },
          "end_index": {
            "type": "integer",
            "exclusiveMinimum": 0.0,
            "title": "End Index"
          },
          "entity_type": {
            "type": "string",
            "title": "Entity Type",
            "default": "MANUAL_SELECTION"
          },
          "entity_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Entity Name"
          },
          "transformation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TransformationRule"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "start_index",
          "end_index"
        ],
        "title": "ManualTextSpan",
        "description": "A manually selected range in text."
      },
...
      "PaginationMeta": {
        "properties": {
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          },
          "total_items": {
            "type": "integer",
            "title": "Total Items"
          },
          "total_pages": {
            "type": "integer",
            "title": "Total Pages"
          },
          "has_next": {
            "type": "boolean",
            "title": "Has Next"
          },
          "has_previous": {
            "type": "boolean",
            "title": "Has Previous"
          }
        },
        "type": "object",
        "required": [
          "page",
          "page_size",
          "total_items",
          "total_pages",
          "has_next",
          "has_previous"
        ],
        "title": "PaginationMeta",
        "description": "Pagination metadata returned alongside paged results."
      },
...
      "ReplacementRecord": {
        "properties": {
          "entity_type": {
            "type": "string",
            "title": "Entity Type"
          },
          "start_index": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Index"
          },
          "end_index": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Index"
          },
          "original_preview": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Original Preview"
          },
          "output_value": {
            "type": "string",
            "title": "Output Value"
          }
        },
        "type": "object",
        "required": [
          "entity_type",
          "output_value"
        ],
        "title": "ReplacementRecord",
        "description": "One applied replacement in a transformed output."
      },
      "ReviewDecision": {
        "type": "string",
        "enum": [
          "pending",
          "approved",
          "rejected"
        ],
        "title": "ReviewDecision"
      },
...
      "TextAnalysisRequest": {
        "properties": {
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "content": {
            "type": "string",
            "title": "Content"
          },
          "content_type": {
            "$ref": "#/components/schemas/ContentType",
            "default": "text"
          },
          "apply_builtins": {
            "type": "boolean",
            "title": "Apply Builtins",
            "default": true
          },
          "pattern_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Pattern Ids"
          },
          "custom_entity_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Custom Entity Ids"
          },
          "configuration_ids": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Configuration Ids"
          },
          "exact_values": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Exact Values"
          },
          "manual_spans": {
            "items": {
              "$ref": "#/components/schemas/ManualTextSpan"
            },
            "type": "array",
            "title": "Manual Spans"
          },
          "default_transformation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TransformationRule"
              },
              {
                "type": "null"
              }
            ]
          },
          "persist_job": {
            "type": "boolean",
            "title": "Persist Job",
            "default": true
          },
          "persist_source_content": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Persist Source Content"
          }
        },
        "type": "object",
        "required": [
          "content"
        ],
        "title": "TextAnalysisRequest",
        "description": "Request to detect findings in text-like content."
      },
      "TextAnalysisResponse": {
        "properties": {
          "job_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Id"
          },
          "findings": {
            "items": {
              "$ref": "#/components/schemas/FindingRecord"
            },
            "type": "array",
            "title": "Findings"
          },
          "summary": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Summary"
          }
        },
        "type": "object",
        "required": [
          "findings",
          "summary"
        ],
        "title": "TextAnalysisResponse",
        "description": "Detection result for text workflows."
      },
      "TextTransformRequest": {
        "properties": {
          "content": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content"
          },
          "job_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Id"
          },
          "finding_overrides": {
            "items": {
              "$ref": "#/components/schemas/FindingOverride"
            },
            "type": "array",
            "title": "Finding Overrides"
          },
          "include_pending": {
            "type": "boolean",
            "title": "Include Pending",
            "default": false
          },
          "default_transformation": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TransformationRule"
              },
              {
                "type": "null"
              }
            ]
          },
          "persist_output": {
            "type": "boolean",
            "title": "Persist Output",
            "default": true
          }
        },
        "type": "object",
        "title": "TextTransformRequest",
        "description": "Request to transform text from findings."
      },
      "TextTransformResponse": {
        "properties": {
          "job_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Job Id"
          },
          "output_text": {
            "type": "string",
            "title": "Output Text"
          },
          "replacements": {
            "items": {
              "$ref": "#/components/schemas/ReplacementRecord"
            },
            "type": "array",
            "title": "Replacements"
          },
          "summary": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Summary"
          }
        },
        "type": "object",
        "required": [
          "output_text",
          "replacements",
          "summary"
        ],
        "title": "TextTransformResponse",
        "description": "Transformation result for text content."
      },
      "TransformationMode": {
        "type": "string",
        "enum": [
          "generic",
          "semantic",
          "custom",
          "partial_mask",
          "stable_alias",
          "mask",
          "blur",
          "pixelate",
          "overlay",
          "image_replacement"
        ],
        "title": "TransformationMode"
      },
      "TransformationRule": {
        "properties": {
          "mode": {
            "$ref": "#/components/schemas/TransformationMode",
            "default": "generic"
          },
          "placeholder": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Placeholder"
          },
          "semantic_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Semantic Label"
          },
          "alias_prefix": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Alias Prefix"
          },
          "prefix_visible": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Prefix Visible",
            "default": 0
          },
          "suffix_visible": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Suffix Visible",
            "default": 0
          },
          "mask_character": {
            "type": "string",
            "title": "Mask Character",
            "default": "*"
          },
          "blur_radius": {
            "type": "integer",
            "minimum": 1.0,
            "title": "Blur Radius",
            "default": 8
          },
          "pixelation_scale": {
            "type": "integer",
            "minimum": 2.0,
            "title": "Pixelation Scale",
            "default": 8
          },
          "overlay_color": {
            "type": "string",
            "title": "Overlay Color",
            "default": "#111111"
          },
          "overlay_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Overlay Label"
          }
        },
        "type": "object",
        "title": "TransformationRule",
        "description": "How a finding should be transformed for output."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  },
...
}